FIX:Change BSIZE from i32 to usize

+++ kernel-rs/src/file.rs
@@ -218,7 +218,7 @@ impl File {
             // and 2 blocks of slop for non-aligned writes.
             // this really belongs lower down, since write()
             // might be writing a device like the console.
-            let max = (MAXOPBLOCKS - 1 - 1 - 2) / 2 * BSIZE;
+            let max = (MAXOPBLOCKS - 1 - 1 - 2) / 2 * (BSIZE as i32);
             let mut i: i32 = 0;
             while i < n {
                 let mut n1: i32 = n - i;
